Refactor extras
