+++ src/game/data/spriteType.ts
@@ -1,8 +1,9 @@
 
-export const SpriteType = {
-    LIFE: 4,
-    MAGIC: 5,
-    KEY: 6,
-    KASHES: 18,
-    ZLITOS: 19,
-};
+export enum SpriteType {
+    UNKNOWN = 0,
+    LIFE = 4,
+    MAGIC = 5,
+    KEY = 6,
+    KASHES = 18,
+    ZLITOS = 19
+}
